Item12952: jslinting resig
remaining errors:

jslint:resig.uncompressed.js:10:62:'xyz' is not defined.
jslint:resig.uncompressed.js:10:62:Expected an assignment or function call and instead saw an expression.
jslint:resig.uncompressed.js:45:10:Be careful when making functions within a loop. Consider putting the function in a closure.
jslint:resig.uncompressed.js:26:9:The body of a for in should be wrapped in an if statement to filter unwanted properties from the prototype.